Centralite Keypad reporting 0% battery after Groovy update
The recent Groovy update changed how the following line of code works:

    result.value = Math.min(100, (int) pct * 100)

Previously the multiplication would happen before the typecast so it would
yield the desired result. With the new version of Groovy that's now used
to run DTHs the typecast happens before the multiplication, so any value of
pct between 0 and 1 would be truncated to 0 resulting in a battery percentage
of 0%. This change fixes that issue, plus rounds to the nearest percentage
instead of truncating and uses a minimum percentage of 1 instead of 0. The
reasoning behind that is if the battery message is received at all then it
has at least a little bit of juice left.